feat: fix flag type

+++ cmd/vegawallet/commands/service_run.go
@@ -129,7 +129,7 @@ func BuildCmdRunService(w io.Writer, handler RunServiceHandler, rf *RootFlags) *
 		"",
 		"Path to the file containing the tokens database passphrase",
 	)
-	cmd.PersistentFlags().Int64Var(&f.TTL,
+	cmd.PersistentFlags().Uint64Var(&f.TTL,
 		"transaction-ttl",
 		0,
 		"Specify number of blocks during which transactions are considered current",
@@ -147,7 +147,7 @@ type RunServiceFlags struct {
 	TokensPassphraseFile   string
 	NoVersionCheck         bool
 	tokensPassphrase       string
-	TTL                    int64
+	TTL                    uint64
 }
 
 func (f *RunServiceFlags) Validate(rf *RootFlags) error {
@@ -285,7 +285,7 @@ func RunService(w io.Writer, rf *RootFlags, f *RunServiceFlags) error {
 
 	jobRunner := vgjob.NewRunner(context.Background())
 
-	svcURL, errChan, err := serviceStarter.Start(jobRunner, f.Network, f.NoVersionCheck, uint64(f.TTL))
+	svcURL, errChan, err := serviceStarter.Start(jobRunner, f.Network, f.NoVersionCheck, f.TTL)
 	if err != nil {
 		cliLog.Error("Failed to start HTTP server", zap.Error(err))
 		jobRunner.StopAllJobs()
